+++ src/settings/baseWidget.schema.ts
@@ -1,4 +1,12 @@
-import { ColorSchema, EFontWeight, EWidgetFilterMode, themed, type TZod } from "..";
+import {
+  ColorSchema,
+  EFontWeight,
+  EWidgetFilterMode,
+  themed,
+  type TZod,
+  VersionedEnum,
+  zodEnumLike,
+} from "..";
 import { ActionButtonSchema } from "../actions.schema";
 import { SettingsFilterSchema } from "../filtration.schema";
 import { MarkdownMeasureSchema, WidgetSortingIndicatorSchema } from "../indicators.schema";
@@ -34,13 +42,16 @@ export const BaseWidgetSettingsSchema = SchemaRegistry.define({
   key: "BaseWidgetSettings",
   latestVersion: "19",
   get history() {
-    const v17 = (z: TZod) =>
-      z.object({
+    const v17 = (z: TZod) => {
+      const v17FilterMode = EWidgetFilterMode[VersionedEnum.forVersion]("17");
+      const v17FontWeight = EFontWeight[VersionedEnum.forVersion]("17");
+
+      return z.object({
         title: z.string().default(""),
         titleSize: themed(z.number().default(14), (theme) => theme.widgets.titleSize),
         titleColor: themed(ColorSchema.forVersion("17")(z), (theme) => theme.widgets.titleColor),
         titleWeight: themed(
-          z.enum(EFontWeight).default(EFontWeight.NORMAL),
+          z.enum(zodEnumLike(v17FontWeight)).default(v17FontWeight.NORMAL),
           (theme) => theme.widgets.titleWeight
         ),
         stateName: z.string().nullable().default(null),
@@ -49,17 +60,23 @@ export const BaseWidgetSettingsSchema = SchemaRegistry.define({
         markdownText: z.string().default(""),
         markdownTextSize: z.number().default(14),
         filters: z.array(SettingsFilterSchema.forVersion("17")(z)).default([]),
-        filterMode: z.enum(EWidgetFilterMode).default(EWidgetFilterMode.DEFAULT),
+        filterMode: z.enum(zodEnumLike(v17FilterMode)).default(v17FilterMode.DEFAULT),
         ignoreFilters: z.boolean().default(false),
         sorting: z.array(WidgetSortingIndicatorSchema.forVersion("17")(z)).default([]),
         actionButtons: z.array(ActionButtonSchema.forVersion("17")(z)).default([]),
         paddings: z.union([z.number(), z.string()]).default(8),
         viewTheme: z.boolean().default(false),
       });
+    };
 
     return {
       "17": v17,
-      "19": (z: TZod) => v17(z).omit({ paddings: true }),
+      "19": (z: TZod) =>
+        v17(z)
+          .omit({ paddings: true, filterMode: true })
+          .extend({
+            filterMode: z.enum(zodEnumLike(EWidgetFilterMode)).default(EWidgetFilterMode.DEFAULT),
+          }),
     };
   },
 });